[modify] update, delete 시 user의 게시글이 맞는지 확인. 유저 인증 코드 수정

+++ core/domains/board/use_case/v1/post_use_case.py
@@ -11,15 +11,23 @@ from core.use_case_output import UseCaseSuccessOutput, UseCaseFailureOutput, Fai
 
 
 class PostBaseUseCase:
+    @inject.autoparams()
+    def __init__(self, board_repo: BoardRepository):
+        self._board_repo = board_repo
+
     def _get_user(self, user_id: int) -> Optional[UserEntity]:
         send_message(UserTopicEnum.GET_USER, user_id=user_id)
 
         return get_event_object(UserTopicEnum.GET_USER)
 
+    def _is_post_owner(self, dto):
+        return self._board_repo.is_post_owner(dto=dto)
+
 
 class CreatePostUseCase(PostBaseUseCase):
     @inject.autoparams()
     def __init__(self, board_repo: BoardRepository):
+        super().__init__(board_repo)
         self._board_repo = board_repo
 
     def execute(
@@ -38,11 +46,16 @@ class CreatePostUseCase(PostBaseUseCase):
 class UpdatePostUseCase(PostBaseUseCase):
     @inject.autoparams()
     def __init__(self, board_repo: BoardRepository):
+        super().__init__(board_repo)
         self._board_repo = board_repo
 
     def execute(
         self, dto: UpdatePostDto
     ) -> Union[UseCaseSuccessOutput, UseCaseFailureOutput]:
+        is_post_owner = self._is_post_owner(dto=dto)
+        if not is_post_owner:
+            return UseCaseFailureOutput(type=FailureType.INVALID_REQUEST_ERROR)
+
         post = self._board_repo.update_post(dto=dto)
         if not post:
             return UseCaseFailureOutput(type=FailureType.SYSTEM_ERROR)
@@ -52,11 +65,16 @@ class UpdatePostUseCase(PostBaseUseCase):
 class DeletePostUseCase(PostBaseUseCase):
     @inject.autoparams()
     def __init__(self, board_repo: BoardRepository):
+        super().__init__(board_repo)
         self._board_repo = board_repo
 
     def execute(
         self, dto: DeletePostDto
     ) -> Union[UseCaseSuccessOutput, UseCaseFailureOutput]:
+        is_post_owner = self._is_post_owner(dto=dto)
+        if not is_post_owner:
+            return UseCaseFailureOutput(type=FailureType.INVALID_REQUEST_ERROR)
+
         post = self._board_repo.delete_post(dto=dto)
         if not post:
             return UseCaseFailureOutput(type=FailureType.SYSTEM_ERROR)
